Move the classes to the outer module
