+++ lib/docker/dockerrun.rb
@@ -1,0 +1,5 @@
+module Docker
+  module Dockerrun
+    FILE_NAME = 'Dockerrun.aws.json'
+  end
+end
